+++ craft_application/services/fetch.py
@@ -85,9 +85,10 @@ class FetchService(services.ProjectService):
 
         if not self._services.ProviderClass.is_managed():
             # Emit a warning, but only on the host-side.
+            logpath = fetch.get_log_filepath()
             emit.message(
-                "Warning: the fetch-service integration is experimental "
-                "and still in development."
+                "Warning: the fetch-service integration is experimental. "
+                f"Logging output to {str(logpath)!r}."
             )
 
             self._fetch_process = fetch.start_service()