+++ tests/integration/services/test_fetch.py
@@ -174,7 +174,13 @@ def test_service_logging(app_service, mocker, tmp_path, monkeypatch, mock_instan
     monkeypatch.chdir(tmp_path)
     mocker.patch.object(fetch, "_get_gateway", return_value="127.0.0.1")
 
-    logfile = fetch._get_log_filepath()
+    # Mock get_log_filepath() so that the test doesn't interfere with whatever
+    # fetch-service is running on the system.
+    mocker.patch.object(
+        fetch, "get_log_filepath", return_value=tmp_path / "craft-fetch-log.txt"
+    )
+
+    logfile = fetch.get_log_filepath()
     assert not logfile.is_file()
 
     app_service.setup()